+++ src/main/java/ru/egar/myOrg/worker/repository/EmployPositionRepository.java
@@ -3,5 +3,5 @@ package ru.egar.myOrg.worker.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.egar.myOrg.worker.model.EmployPosition;
 
-public interface EmployPositionRepository extends JpaRepository<EmployPosition, Integer> {
+public interface EmployPositionRepository extends JpaRepository<EmployPosition, Long> {
 }
